Merged PR 18: Updated textbox delete, fixed message not being included in email, removed href attributes form sidebar buttons
User can now delete a textbox with the delete key or ctrl + q
Fixed the user's email message not being included in the email.
Removed the href attributes from the links in the sidebar, so that they don't redirect.
Removed unused sidebar buttons, so it's not cluttered.
Removed console.log() statements on the client side.

Related work items: #33